Mudando a tipagem das variáveis

+++ Program.cs
@@ -1,17 +1,17 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-int num1;
-int num2;
+double num1;
+double num2;
 Console.WriteLine("Digite o valor do primeiro número: ");
-num1 = Convert.ToInt32(Console.ReadLine());
+num1 = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Perfeito! Agora digite o valor do segundo número:");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 
-int soma = num1 + num2;
-int subtracao = num1 - num2;
-int multiplicacao = num1 * num2;
-double divisao = (double) num1 / num2;
+double soma = num1 + num2;
+double subtracao = num1 - num2;
+double multiplicacao = num1 * num2;
+double divisao = num1 / num2;
 //Formatação de String utilizando Concatenação!
 Console.WriteLine("A soma dos dois números é:" + soma);
 //Formatação de String utilizando Interpolação!
